requirements.txt
aiohttp
aiosignal
altair<5
annotated-types
anyio
async-timeout
attrs
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
dataclasses-json
distro
entrypoints
exceptiongroup
Faker
favicon
fonttools
frozenlist
gitdb
GitPython
greenlet
h11
htbuilder
httpcore
httpx
idna
importlib-metadata
importlib_resources
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
lxml
Markdown
markdown-it-py
markdownlit
MarkupSafe
marshmallow
matplotlib
mdurl
more-itertools
multidict
mypy-extensions
numpy
openai
orjson
packaging
pandas
Pillow
prometheus_client
protobuf
pyarrow
pycryptodome
pydantic
pydantic_core
pydeck
Pygments
pymdown-extensions
Pympler
pyparsing
python-dateutil
python-dotenv
pytz
pytz-deprecation-shim
PyYAML
referencing
regex
requests
rich
rpds-py
six
smmap
sniffio
soupsieve
SQLAlchemy
st-annotated-text
streamlit-camera-input-live
streamlit-card
streamlit-embedcode
streamlit-extras
streamlit-faker
streamlit-image-coordinates
streamlit-keyup
streamlit-toggle-switch
streamlit-vertical-slider
tenacity
tiktoken
toml
toolz
tornado
tqdm
typing-inspect
typing_extensions
tzdata
tzlocal
urllib3
validators
watchdog
yarl
zipp
langchain==0.0.154
PyPDF2==3.0.1
streamlit==1.18.1
faiss-cpu==1.7.4
googletrans